pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "FastAPI_Azure_auth"
version = "0.0.1"
description = "FastApi Azure Authentication"
classifiers = ["Private :: Do Not Upload", "Programming Language :: Python :: 3.11"]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "starlette",
    "fastapi==0.109.1",
    "uvicorn==0.23.2",
    "environs==9.5.0",
    "itsdangerous",
    "apscheduler",
    "asyncpg",
    "httpx",
    "pydantic",
    "pydantic[email]",
    "passlib",
    "python-multipart",
    "bcrypt==4.0.1",
    "python-jose"
]

[project.urls]
Source = "https://github.com/ktsoftuni/fastapi_azure_auth"
